Improved the TextField object and fixed some features - Implemented the Fonts pooling - Created the CustomFont struct for an easier management - Adjusted the TextField size - Implemented the transparent background and color on TextFields - Started the rotation reworking with floats

+++ PathfindingAlgorithmsProject/PathfindingAlgorithmsProject/GameObject.cpp
@@ -133,9 +133,9 @@ short GameObject::GlobalPivotY() const
 short GameObject::GlobalRotation() const
 {
 	if (_parent == nullptr)
-		return rotation % 360;
+		return (short)rotation % 360;
 	else
-		return (_parent->GlobalRotation() + rotation) % 360;
+		return (short)(_parent->GlobalRotation() + rotation) % 360;
 }
 
 float GameObject::GlobalA() const
